Fix libretro
