+++ CMakeLists.txt
@@ -16,8 +16,10 @@ if ("${CMAKE_SOURCE_DIR}" STREQUAL "${CMAKE_CURRENT_SOURCE_DIR}")
 else()
     set(PNTR_APP_IS_MAIN FALSE)
 endif()
-option(PNTR_APP_BUILD_EXAMPLES "Build Examples" ${PNTR_APP_IS_MAIN})
+option(PNTR_BUILD_EXAMPLE_SDL "Example: SDL" ${PNTR_APP_IS_MAIN})
+option(PNTR_BUILD_EXAMPLE_RAYLIB "Example: raylib" ${PNTR_APP_IS_MAIN})
+option(PNTR_BUILD_EXAMPLE_LIBRETRO "Example: libretro" ${PNTR_APP_IS_MAIN})
 
-if (PNTR_APP_BUILD_EXAMPLES)
-    add_subdirectory(examples)
+if (PNTR_BUILD_EXAMPLE_SDL OR PNTR_BUILD_EXAMPLE_RAYLIB OR PNTR_BUILD_EXAMPLE_LIBRETRO)
+    add_subdirectory(example)
 endif()
