+++ cmake/FindSDL2.cmake
@@ -1,3 +1,4 @@
+# SDL2
 include(FetchContent)
 FetchContent_Declare(
     SDL2Source
