+++ cmake/Findraylib.cmake
@@ -1,3 +1,4 @@
+# raylib
 include(FetchContent)
 FetchContent_Declare(
     raylib
