+++ example/pntr_app_example.c
@@ -1,6 +1,6 @@
 #define PNTR_APP_IMPLEMENTATION
 #define PNTR_DISABLE_MATH
-#include "pntr_app.h"
+#include "../include/pntr_app.h"
 
 bool Init(void* userData) {
     // Nothing.
@@ -9,7 +9,7 @@ bool Init(void* userData) {
 }
 
 bool Update(pntr_image* screen, void* userData) {
-    pntr_clear_background(screen, PNTR_WHITE);
+    pntr_clear_background(screen, PNTR_RAYWHITE);
     pntr_draw_circle_fill(screen, screen->width / 2, screen->height / 2, 100, PNTR_BLUE);
 
     return true;
@@ -21,8 +21,8 @@ void Close(void* userData) {
 
 pntr_app Main(int argc, char* argv[]) {
     return (pntr_app) {
-        .width = 800,
-        .height = 450,
+        .width = 400,
+        .height = 225,
         .title = "pntr_app",
         .init = Init,
         .update = Update,
