+++ example/resources/file.txt
